refactor(direct-control): extract tuner framing helpers
Move Civ7TunerFrame plus encodeCiv7TunerRequest and parseCiv7TunerFrame into packages/civ7-direct-control/src/session/framing.ts behind the existing package facade. Keep broader session/config/reconnect and restart/setup lifecycle source extraction pending.

Record queued source-agent dispositions before closing this slice: the primitives candidate was partially accepted only for the already committed ComponentID/error subset, the framing candidate was integrated on current disk, and the read-only atom lane is planning evidence for a later settlementRecommendationsSource move.

Validation: bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts restart-lifecycle.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
